requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
discord==1.0.1
discord.py==1.0.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
peewee==3.9.5
pickleDB==0.9.2
#-e git+git@github.com:python-kol/pykollib.git@15aa8b8d1e0e4ea6c5ad8b955e79e8f05dbef66c#egg=pykollib
python-dotenv==0.10.1
requests==2.21.0
typing-extensions==3.7.2
urllib3==1.24.2
websockets==6.0
yarl==1.3.0
tqdm==4.32.2
tabulate==0.8.3